Refactor: 관심사 분리
