+++ src/hooks/useAuth.js
@@ -1,9 +1,12 @@
 import { useState } from "react";
-import { signin } from "../apis/auth";
+import { signin, signup } from "../apis/auth";
+import { useNavigate } from "react-router-dom";
 
 const useAuth = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+
+  const navigate = useNavigate();
 
   const handleEmailChange = (event) => {
     setEmail(event.target.value);
@@ -13,7 +16,7 @@ const useAuth = () => {
     setPassword(event.target.value);
   };
 
-  const handleSignin = (navigate) => {
+  const handleSignin = () => {
     signin(email, password)
       .then((res) => {
         console.log(res);
@@ -27,12 +30,26 @@ const useAuth = () => {
       });
   };
 
+  const handleSignup = () => {
+    signup(email, password)
+      .then((res) => {
+        console.log(res);
+        alert("회원가입 완료");
+        navigate("/signin");
+      })
+      .catch((err) => {
+        console.log(err);
+        alert("회원가입 실패");
+      });
+  };
+
   return {
     email,
     password,
     handleEmailChange,
     handlePasswordChange,
     handleSignin,
+    handleSignup,
   };
 };
